refactor(core): split label formatters and M2 pipeline characterization
Extract format-time, format-temporal-labels, and format-number from the
mixed layout/format module; re-export the public surface from format.ts.

Replace the 539-line pipeline-m2 monolith with focused suites under
tests/pipeline-m2/ (histogram, smooth, boxplot, density, errorbar,
positions, render-svg) sharing scatter fixtures.
